+++ packages/ricos-modals/web/src/index.ts
@@ -2,4 +2,5 @@ export { RicosModalService } from './ModalService';
 export { ModalContext } from './utils/ModalContext';
 export { ModalProvider } from './utils/ModalProvider';
 export { withModalContext } from './utils/withModalContext';
+export { PLACEMENTS, LAYOUTS } from './consts';
 export type { ModalContextValue } from './types';